pyproject.toml
[build-system]
#######################
# Python build system #
#######################
# The build system to use when installing this package.
# Used when installing the package with `pip install .`.
# See also: https://www.python.org/dev/peps/pep-0517/

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.poetry]
###################
# Poetry settings #
###################
# See https://python-poetry.org/docs/pyproject/ for more details!

# The name of your project.
# Ensure that it is available on PyPI: https://pypi.org/
name = "temple_of_styx"

# The version of the package.
version = "0.1.0"

# A brief, one-sentence description about your project.
description = "Authentication and authorization server for Star Shard Studio"

# A list of the authors of the project.
authors = [
    "Stefano Pigozzi <me@steffo.eu>",
]

# A list of maintainers of the project.
# Often, it is the same as the authors list.
maintainers = [
    "Stefano Pigozzi <me@steffo.eu>",
]

# The license of the package.
# Uses SPDX format: https://spdx.org/licenses/
license = "AGPL-3.0-or-later"

# The README file.
readme = "README.md"

# The URL of the project website.
# Not the GitHub repository!
#homepage = "https://example.org/"

# The URL of the project repository.
repository = "https://github.com/starshardstudio/temple-of-styx/"

# The URL of the project documentation location.
#documentation = "https://temple-of-styx.readthedocs.io/latest/"

# Up to five keywords related to your project.
# See also: https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#keywords
#keywords = [
#    "",
#    "",
#    "",
#    "",
#    "",
#]

# Any number of trove classifiers that apply to your project.
# See the list at: https://pypi.org/classifiers/
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]

# ADVANCED: specify the packages exported by your project
# See also: https://python-poetry.org/docs/pyproject/#packages
# packages = [
#     # Regular packages
#     { include = "OTHER_PACKAGE_NAME" },
#     # Namespace packages
#     { include = "EXTEND/OTHER/NAMESPACE/**/*.py" }
# ]

# ADVANCED: include additional files in the build
# include = [
#     "FILENAME.txt"
# ]

# ADVANCED: exclude files from the build
# exclude = [
#     "PACKAGE_NAME/test.py"
# ]



[tool.poetry.urls]
##################
# URLs           #
##################
# Additional project URLs in a name → link mapping.



[tool.poetry.scripts]
#####################
# Scripts           #
#####################
# Poetry can create "binaries" to add to the PATH when the project is installed.
# They are specified in the form of a mapping with the command name as key and the function to execute as value.

styxctl = "temple_of_styx.cli.__main__:main"



[tool.poetry.dependencies]
##########################
# Dependencies           #
##########################
# A mapping of dependencies of the project
# It should be maintained by `poetry add` / `poetry remove`, but it currently adds things after all comments...
# You can manually specify allowed version numbers:
# * means "any release"
#         *        →  any
# ^X.X.X means "newer releases with this major version"
#         ^3.10.1  →  == 3      && >= 3.10.1
# ~X.X.X means "newer releases with this minor version"
#         ~3.10.1  →  == 3.10   && >= 3.10.1
# nothing means "this specific release"
#          3.10.1  →  == 3.10.1

python = "^3.10"
flask = "^2.2.3"
authlib = "^1.2.0"
flask-sqlalchemy = "^3.0.3"
psycopg2 = "^2.9.6"
alembic = "^1.10.3"
cfig = {extras = ["cli"], version = "^0.3.0"}
argon2-cffi = "^21.3.0"
flask-login = "^0.6.2"
click = "^8.1.3"
webauthn = "^2.1.0"



[tool.poetry.extras]
####################
# Package extras   #
####################
# ADVANCED: specify optional dependency groups.
# See: https://python-poetry.org/docs/pyproject/#extras



[tool.poetry.plugins]
#####################
# Poetry plugins    #
#####################
# ADVANCED: extend Poetry's functionality.
# See: https://python-poetry.org/docs/pyproject/#plugins



[tool.pytest.ini_options]
#########################
# Pytest configuration  #
#########################
